Refactor feature tests to separate class.
